+++ zxq/pipeline/steps/__init__.py
@@ -2,10 +2,12 @@ from .data_cleanser import DataCleanser
 from .feature_engineer import FeatureEngineer
 from .missing_value_handler import MissingValueHandler
 from .time_aligner import TimeAligner
+from .schema_validator import SchemaValidatorStep
 
 __all__ = [
     "DataCleanser",
     "FeatureEngineer",
     "MissingValueHandler",
     "TimeAligner",
+    "SchemaValidatorStep",
 ]